Replace the macro-generated encoders with generic ones
`core/src/encoders.rs` defined 33 encoder structs and 33 builder structs, almost
all emitted by four near-identical macros (`impl_encode!`,
`impl_encode_fallible!`, `impl_encoder_builder_stateless!`,
`impl_encoder_builder_stateless_with_field!`). What actually varied between them
was a triple: which Arrow array to read, which Postgres type to declare, and how
to write one value.

That triple is now a *conversion* -- a zero-sized marker type implementing
`FixedSizeConversion` or `NumericConversion` -- and the encoders are generic
over it. The 66 hand-rolled types collapse to 6 encoders and 7 builders:

  FixedSizeEncoder<C>   / FixedSizeEncoderBuilder<C>   21 conversions
  NumericEncoder<C>     / NumericEncoderBuilder<C>     4 conversions (3 decimal
                                                      widths + UInt64)
  GenericBinaryEncoder  / GenericBinaryEncoderBuilder<T: OffsetSizeTrait>
  GenericStrEncoder     / StrEncoderBuilder<C>
  GenericListEncoder    / GenericListEncoderBuilder<T: OffsetSizeTrait>
  StructEncoder         / StructEncoderBuilder
                        / Int8EncoderBuilder (the one caller-selectable output)

All four macros are gone, and with them the `#![allow(clippy::redundant_closure_call)]`
that only existed to serve the macro style. The remaining macro,
`encode_decimal!`, emits three free functions over i32/i64/i128 and is about
integer-width polymorphism, not boilerplate.

The public API stays source-compatible: every `FooEncoder` / `FooEncoderBuilder`
is now a type alias for the corresponding generic instantiation, so
`StringEncoderBuilder::new_with_output`, `ListEncoderBuilder::new_with_inner`,
`BuildEncoder`, `schema()` and all 33 `EncoderBuilder` / `Encoder` variants keep
working unchanged -- which is what `py/src` depends on.

The file is split into `core/src/encoders/`:

  mod.rs     traits, the two dispatch enums, the aliases, `try_new`
  scalar.rs  the two generic scalar encoders and the conversion table
  numeric.rs the NUMERIC wire format and its unit tests
  text.rs    bytea and the three string layouts
  nested.rs  lists and structs

Two mechanisms are worth calling out. A conversion's field size is derived from
its Postgres type (`PostgresType::size()`) instead of being restated, because
they agreed for all 21 fixed-size encoders anyway; it is resolved once per batch
rather than per row. And a builder produces the right `Encoder` variant through
the `From` impls `enum_dispatch` already generates, via a
`for<'a> FixedSizeEncoder<'a, C>: Into<Encoder<'a>>` bound, so there is no
per-type wrapping function.

Behaviour is unchanged except for one latent bug that the generic form fixes:
`LargeListEncoderBuilder::new` matched `DataType::List` rather than
`DataType::LargeList` (the macro hardcoded `List` for both instantiations), so
it rejected the very fields it exists for and accepted ones it would then fail
to downcast. Nothing in-tree called it. `EncoderBuilder::try_new` keeps its
existing lenient handling of `FixedSizeBinary`/`FixedSizeList` by constructing
builders through an unchecked path.

Wire format is byte-identical: `tests/snapshots/*.bin` and `snapshots_csv/*` are
untouched and `validate_snapshots` / `validate_roundtrip_values` /
`decimal_wire_format` / the proptest suite all pass.

Addresses items 1 and part of 6 of #69.
